finalproject-draft1
hi! the css/html structure is a little bare right now, as I want the
site to be minimal, with a few v. important interactive components.
I’ve already started making notes in the js file included here, and as
i continue working on this, tweaking the organization of content and
finding what i want it to look/respond like, it will likely change
quite a bit. There are some v. specific things that I don;t know how to
do that I will be asking questions about :D